% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{analyze_conversions_after_index}
\alias{analyze_conversions_after_index}
\title{Analyze Conversions After Index}
\usage{
analyze_conversions_after_index(
  convert_data,
  seqs,
  seq2pt,
  ip_seqs,
  ip_pt_seqs,
  dates,
  pt_trace
)
}
\arguments{
\item{convert_data}{Output from analyze_convert_patients()}

\item{seqs}{Vector of sequence IDs in the cluster}

\item{seq2pt}{Named vector mapping sequence IDs to patient IDs}

\item{ip_seqs}{Vector of sequence IDs corresponding to imported cases}

\item{ip_pt_seqs}{Vector of sequence IDs corresponding to intake positive patients}

\item{dates}{Vector of isolate dates named by sequence IDs}

\item{pt_trace}{Matrix with rows representing days and columns representing patients}
}
\value{
Vector with conversion timing properties
}
\description{
Analyze conversions relative to index patient timing
}
\keyword{internal}
