% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{analyze_convert_patients}
\alias{analyze_convert_patients}
\title{Analyze Convert Patients}
\usage{
analyze_convert_patients(seqs, seq2pt, ip_pt_seqs, dates, pt_trace)
}
\arguments{
\item{seqs}{Vector of sequence IDs in the cluster}

\item{seq2pt}{Named vector mapping sequence IDs to patient IDs}

\item{ip_pt_seqs}{Vector of sequence IDs corresponding to intake positive patients}

\item{dates}{Vector of isolate dates named by sequence IDs}

\item{pt_trace}{Matrix with rows representing days and columns representing patients}
}
\value{
List containing convert patient analysis results
}
\description{
Analyze patients who converted while in facility
}
\keyword{internal}
