% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_epi_stats.R
\name{analyze_overlap_by_cluster}
\alias{analyze_overlap_by_cluster}
\title{Analyze Overlap by Cluster}
\usage{
analyze_overlap_by_cluster(clusters, trace_mat)
}
\arguments{
\item{clusters}{A named numeric vector where names are sequence IDs and values are subtrees defining the cluster.}

\item{trace_mat}{A matrix with rows representing days and columns representing isolates}
}
\value{
A data frame with the overlap between isolates
}
\description{
Analyze the overlap between isolates in a trace matrix by cluster
}
