% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_epi_stats.R
\name{calculate_trace_overlap}
\alias{calculate_trace_overlap}
\title{Calculate Trace Overlap}
\usage{
calculate_trace_overlap(trace_mat)
}
\arguments{
\item{trace_mat}{A matrix with rows representing days and columns representing isolates}
}
\value{
A matrix with the overlap between isolates
}
\description{
Calculate the overlap between isolates in a trace matrix
}
