% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_categorize.R
\name{categorize_cluster_overlap}
\alias{categorize_cluster_overlap}
\title{Categorize cluster overlap}
\usage{
categorize_cluster_overlap(isolate_lookup, cluster_overlap_df, surv_df)
}
\arguments{
\item{isolate_lookup}{A lookup table for isolates and their clusters assignments which has
other relevant epidemiological information. For more information, see
\code{\link[=get_isolate_lookup]{get_isolate_lookup()}}.}

\item{cluster_overlap_df}{A data frame with overlap information for isolate pairs. For more information, see
\code{\link[=cluster_isolate_overlap]{cluster_isolate_overlap()}}.}

\item{surv_df}{A data frame with surveillance information for isolates.}
}
\value{
A vector with the categorization of the overlap for each cluster.
}
\description{
This function categorizes the overlap of isolates in a cluster, based on the earliest collected isolate
from all patients deemed to be in the cluster (hereafter referred to as the "index isolate"), and overlap
explanations for other isolates in the cluster. The categories are:
\itemize{
\item "patient-to-patient": if the index isolate is admission-positive and there is overlap explanation for
all other converts in the cluster.
\item "weak-index": if the index isolate is not admission-positive but is the first surveillance for the patient
after admission and there is overlap explanation for all other isolates in the cluster.
\item "missing-intermediate": if the index isolate is admission-positive but at least one other convert
in the cluster has no overlap explanation.
\item "false-negative-index": if the index isolate is not admission-positive but there is overlap explanation for
all other converts in the cluster barring one (deemed to be the false negative index).
\item "missing-source": if the index isolate is not admission-positive and there is no overlap explanation for
more than one convert in the cluster.
\item "multiply-colonized-index": if the index isolate is not in the cluster but is admission-positive, this is
a "multiply-colonized index" if there is overlap explanation for all other isolates in the cluster.
\item "multiply-colonized-index-missing-intermediate": if the index isolate is not in the cluster but is
admission-positive, and at least one other convert in the cluster has no overlap explanation.
\item "all-admission-positive": if all isolates in the cluster are admission positive, this is a special category
that may indicate a common source or multiple colonization events, but is not explained
\item "inexplicable": catch-all category for cases that do not fit into the other categories. Currently,
index and overlap explanations are not enough to explain these clusters.
}
}
