% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{cluster_context_2}
\alias{cluster_context_2}
\title{Plots the patient trace broken up by SNP threshold and labeled using designated clusters.}
\usage{
cluster_context_2(
  dna_aln,
  pt_trace,
  seq2pt,
  snp_thresh,
  ip_pt_seqs,
  ip_seqs,
  snp_dist,
  dates,
  prefix,
  floor_trace = NULL
)
}
\arguments{
\item{dna_aln}{A DNA object with names corresponding to sequence IDs}

\item{pt_trace}{A data frame or matrix representing daily patient trace (rows) by patient ID (columns)}

\item{seq2pt}{A named vector mapping sequence IDs to patient IDs}

\item{snp_thresh}{Numeric threshold for breaking up trace heatmaps}

\item{ip_pt_seqs}{Vector of sequence IDs for intake positive patients}

\item{ip_seqs}{Vector of sequence IDs presumed to be imported}

\item{snp_dist}{A matrix of SNP distances between isolates}

\item{dates}{A named vector of isolate dates by sequence ID}

\item{prefix}{A string descriptor used for naming figure outputs}

\item{floor_trace}{Optional data for floor-level tracing}
}
\value{
A numeric vector indicating the cluster that each isolate belongs to.
}
\description{
Plots the patient trace broken up by SNP threshold and labeled using designated clusters.
}
\details{
First, we cluster the isolates using a hard SNP distance threshold. Then, we plot
the patient trace heatmaps for each cluster. If a floor trace is provided, we also plot the
floor trace for each cluster.
}
