% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{cluster_genetic_context}
\alias{cluster_genetic_context}
\title{Cluster Genetic Context}
\usage{
cluster_genetic_context(clusters, seq2pt, ip_seqs, snp_dist, prefix)
}
\arguments{
\item{clusters}{A vector named by sequence IDs with values indicating the cluster (subtree) each sequence belongs to.}

\item{seq2pt}{A named vector mapping sequence IDs to patient IDs.}

\item{ip_seqs}{A vector of sequence IDs corresponding to intake positive patients.}

\item{snp_dist}{A matrix of SNP distances between isolates.}

\item{prefix}{A descriptor indicating how the clusters were generated (used to name figure outputs).}
}
\value{
A matrix containing intra-cluster and inter-cluster genetic distances.
}
\description{
Produces summary plots regarding genetic distances within and between transmission clusters.
}
