% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{cluster_intra_versus_inter_context}
\alias{cluster_intra_versus_inter_context}
\title{Plots trace for cluster members and those within the max SNP distance observed within a cluster}
\usage{
cluster_intra_versus_inter_context(
  clusters,
  pt_trace,
  seq2pt,
  ip_seqs,
  dates,
  prefix,
  snp_dist
)
}
\arguments{
\item{clusters}{A vector named by sequence IDs with values being subtrees defining the cluster}

\item{pt_trace}{A trace where the rows are days and the columns are patients}

\item{seq2pt}{A vector of patient IDs named by the corresponding sequence ID}

\item{ip_seqs}{A vector of sequence IDs which correspond to intake positive patients}

\item{dates}{A vector of isolate dates named by sequence IDs}

\item{prefix}{A descriptor of how the clusters were generated for the purposes of naming figures}

\item{snp_dist}{A matrix of SNP distances between isolates}
}
\value{
No return value, invoked for side effects
}
\description{
Plots trace for cluster members and those within the max SNP distance observed within a cluster
}
