% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{cluster_properties}
\alias{cluster_properties}
\title{Compute various cluster-level properties}
\usage{
cluster_properties(
  seqs,
  pt_trace,
  seq2pt,
  ip_pt_seqs,
  ip_seqs,
  snp_dist,
  dates,
  floor_trace = NULL,
  room_trace = NULL
)
}
\arguments{
\item{pt_trace}{A data frame or matrix of patient-level trace data (rows are days,
columns are patient IDs).}

\item{seq2pt}{A named character vector mapping sequence IDs to patient IDs
(names must be sequence IDs, values must be patient IDs).}

\item{ip_pt_seqs}{A character vector of sequence IDs corresponding to patients
who tested positive on admission ("intake positive").}

\item{ip_seqs}{A character vector of sequence IDs presumed to be imported from
outside (a subset of \code{ip_pt_seqs} or similar).}

\item{snp_dist}{A matrix of SNP distances between isolates (row and column
names should match sequence IDs).}

\item{dates}{A named numeric vector of isolate dates by sequence ID.}

\item{floor_trace}{(Optional) A data frame or matrix for floor-level tracing.}

\item{room_trace}{(Optional) A data frame or matrix for room-level tracing.}

\item{cluster_seqs}{A character vector of sequence IDs belonging to the cluster.}
}
\value{
A named numeric vector containing various cluster-level properties. The list
of the properties computed are given in the details section.
}
\description{
Calculates various epidemiological and genetic characteristics for a given
set of sequence IDs in a single cluster. This includes counts of patients,
convert vs. index statuses, genetic distances, cluster duration, and
timing-based metrics such as time to first/last acquisition.
}
\details{
This function computes the following properties for the given cluster:
\itemize{
\item Number_of_patients
\item Number_of_start_indexes
\item Number_of_non-start_indexes
\item Number_of_study_start_indexes
\item Number_of_converts
\item Number_of_initial_converts
\item Number_of_later_converts
\item Mean_genetic_distance
\item Median_genetic_distance
\item Max_genetic_distance
\item Number_of_converts_after_index
\item Number_of_initial_converts_after_index
\item Number_of_converts_after_index_seq
\item Number_of_initial_converts_after_index_seq
\item Number_of_converts_with_source
\item Number_of_initial_converts_with_source
\item Number_of_converts_with_floor_source
\item Number_of_initial_converts_with_floor_source
\item Number_of_converts_with_room_source
\item Number_of_initial_converts_with_room_source
\item Time_to_first_acquisition
\item Time_to_last_acquisition
\item Median_time_to_acquisition
\item Time_from_index_to_first_convert
\item Time_from_index_to_last_convert
}
These can be used for further analysis or visualization, including for example in
the permutation tests to assess the significance of the observed cluster properties.
}
