% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{cluster_property_perm_test}
\alias{cluster_property_perm_test}
\title{Permutation Test for Cluster Properties}
\usage{
cluster_property_perm_test(
  clusters,
  pt_trace,
  seq2pt,
  ip_pt_seqs,
  ip_seqs,
  dates,
  snp_dist,
  prefix,
  nperm = 1000,
  floor_trace = NULL,
  room_trace = NULL,
  mc.cores = detectCores() - 1
)
}
\arguments{
\item{clusters}{A named numeric vector where names are sequence IDs and values are subtrees defining the cluster.}

\item{pt_trace}{A matrix or data frame with rows representing days and columns representing patients.}

\item{seq2pt}{A named vector mapping sequence IDs to patient IDs.}

\item{ip_pt_seqs}{A vector of sequence IDs corresponding to intake positive patients.}

\item{ip_seqs}{A vector of sequence IDs corresponding to intake patient sequences presumed to be imported.}

\item{dates}{A vector of isolate dates named by sequence IDs.}

\item{snp_dist}{A matrix of SNP distances between isolates.}

\item{prefix}{A descriptor string (used to name output figures).}

\item{nperm}{The number of permutations to perform (default is 1000).}

\item{floor_trace}{An optional floor trace (rows: days, columns: patients).}

\item{room_trace}{An optional room trace (rows: days, columns: patients).}

\item{mc.cores}{The number of cores to use for parallel processing (default is one less than the total
number of cores). This is only for *nix systems and does not work on Windows - for Windows,
set mc.cores = 1.}
}
\value{
A matrix of permutation statistics. Rows correspond to each permutation (with the last row
containing the observed statistic) and columns correspond to the computed properties.
}
\description{
This function compares the properties of actual clusters to properties of randomly
permuted clusters with the same size distribution. It computes various summary statistics,
calculates permutation-based p-values, and generates histogram plots of the statistics.
}
