% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{cluster_summary}
\alias{cluster_summary}
\title{Produces summary figures and information regarding the clusters.}
\usage{
cluster_summary(
  clusters,
  sub_trees,
  dna_aln,
  pt_trace,
  seq2pt,
  ip_pt_seqs,
  ip_seqs,
  snp_dist,
  dates,
  prefix,
  floor_trace = NULL
)
}
\arguments{
\item{clusters}{A named vector indicating the cluster (subtree) for each sequence.}

\item{sub_trees}{A list of phylogenetic subtrees.}

\item{dna_aln}{A DNA object corresponding to sequence IDs.}

\item{pt_trace}{A data frame or matrix representing daily patient trace (rows) by patient ID (columns).}

\item{seq2pt}{A named vector mapping sequence IDs to patient IDs.}

\item{ip_pt_seqs}{A vector of sequence IDs which correspond to intake positive patients.}

\item{ip_seqs}{A vector of sequence IDs which correspond to intake positive patients presumed to be imported.}

\item{snp_dist}{A matrix of SNP distances between isolates.}

\item{dates}{A named vector of isolate dates by sequence ID.}

\item{prefix}{A descriptor for naming figure outputs.}

\item{floor_trace}{An optional floor trace where the rows are days and the columns are patients.}
}
\value{
No return value, invoked for side effects.
}
\description{
Produces summary figures and information regarding the clusters.
}
