% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{compare_clusters}
\alias{compare_clusters}
\title{Compares the content of clusters created with two different methods.}
\usage{
compare_clusters(clusters1, clusters2, width = 10, height = 10)
}
\arguments{
\item{clusters1}{A vector named by sequence IDs with values being subtrees defining the cluster.}

\item{clusters2}{A vector named by the same sequence IDs as clusters1 with values being subtrees
defining the cluster.}

\item{width}{The width of the heatmap plot.}

\item{height}{The height of the heatmap plot.}
}
\value{
A heatmap plot object from \code{ComplexHeatmap} showing the overlap between clusters.
}
\description{
Compares the content of clusters created with two different methods.
}
