% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{compute_basic_cluster_counts}
\alias{compute_basic_cluster_counts}
\title{Compute Basic Cluster Counts}
\usage{
compute_basic_cluster_counts(
  seqs,
  seq2pt,
  ip_seqs,
  ip_pt_seqs,
  dates,
  pt_trace
)
}
\arguments{
\item{seqs}{Vector of sequence IDs in the cluster}

\item{seq2pt}{Named vector mapping sequence IDs to patient IDs}

\item{ip_seqs}{Vector of sequence IDs corresponding to imported cases}

\item{ip_pt_seqs}{Vector of sequence IDs corresponding to intake positive patients}

\item{dates}{Vector of isolate dates named by sequence IDs}

\item{pt_trace}{Matrix with rows representing days and columns representing patients}
}
\value{
Vector with basic cluster count properties
}
\description{
Calculate basic patient counts and classifications for a cluster
}
\keyword{internal}
