% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{compute_cluster_duration_metrics}
\alias{compute_cluster_duration_metrics}
\title{Compute Cluster Duration Metrics}
\usage{
compute_cluster_duration_metrics(seqs, seq2pt, dates)
}
\arguments{
\item{seqs}{Vector of sequence IDs in the cluster}

\item{seq2pt}{Named vector mapping sequence IDs to patient IDs}

\item{dates}{Vector of isolate dates named by sequence IDs}
}
\value{
List with cluster duration properties and earliest positive date for each patient
}
\description{
Calculate temporal metrics for cluster spread
}
