% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{compute_duration_metrics_by_cluster}
\alias{compute_duration_metrics_by_cluster}
\title{Compute duration metrics by cluster}
\usage{
compute_duration_metrics_by_cluster(isolate_lookup)
}
\arguments{
\item{isolate_lookup}{An isolate lookup table from \code{\link[=get_isolate_lookup]{get_isolate_lookup()}}.}
}
\value{
Data frame with duration metrics by cluster
}
\description{
Calculate temporal metrics for cluster spread for a set of cluster assignments.
}
