% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{compute_genetic_distances}
\alias{compute_genetic_distances}
\title{Compute Genetic Distances}
\usage{
compute_genetic_distances(seqs, snp_dist)
}
\arguments{
\item{seqs}{Vector of sequence IDs in the cluster}

\item{snp_dist}{Matrix of SNP distances between isolates}
}
\value{
Vector with genetic distance properties
}
\description{
Calculate genetic distance statistics within a cluster
}
\keyword{internal}
