% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{compute_genetic_distances_by_cluster}
\alias{compute_genetic_distances_by_cluster}
\title{Compute genetic distances by cluster}
\usage{
compute_genetic_distances_by_cluster(isolate_lookup, snp_dist)
}
\arguments{
\item{isolate_lookup}{Data frame with isolate lookup information.}

\item{snp_dist}{Matrix of SNP distances between isolates.}
}
\value{
Data frame with genetic distances by cluster
}
\description{
Calculate genetic distances for a set of cluster assignments.
}
