% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeDefiningVariants}
\alias{computeDefiningVariants}
\title{Compute the defining variants for each subtree}
\usage{
computeDefiningVariants(dna_aln, isolate_names, subtrees)
}
\arguments{
\item{dna_aln}{A matrix of DNA sequences (character matrix).}

\item{isolate_names}{A vector of isolate names (character vector).}

\item{subtrees}{A list of subtrees (phylo objects).}
}
\value{
A vector of integers representing the number of defining variants
for each subtree.
}
\description{
Compute the defining variants for each subtree
}
