% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{get_isolate_lookup}
\alias{get_isolate_lookup}
\title{Create a lookup table for isolates along with their cluster assignments.}
\usage{
get_isolate_lookup(clusters, dna_aln, seq2pt, adm_seqs, dates, surv_df)
}
\arguments{
\item{clusters}{A numeric vector of cluster assignments.}

\item{dna_aln}{A DNA alignment object of class \code{DNAbin}.}

\item{seq2pt}{A named vector mapping sequence IDs to patient IDs.}

\item{adm_seqs}{A vector of sequence IDs which correspond to admission positive patient sequences.}

\item{dates}{A named vector mapping sequence IDs to dates.}

\item{surv_df}{A data frame with surveillance information for isolates.}
}
\value{
A data frame with columns: isolate_id, patient_id, date, cluster, adm_pos, prev_surv,
prev_surv_neg.
}
\description{
This function creates a lookup table for isolates. The columns are:
\itemize{
\item cluster: the cluster assignment for the isolate.
\item isolate_id: the isolate ID.
\item patient_id: the patient ID.
\item date: the date of the isolate.
\item adm_pos: a logical value indicating if the isolate is admission positive.
\item prev_surv: the previous surveillance date for the patient.
\item prev_surv_neg: a logical value indicating whether the patient had a prior
surveillance before this isolate and none of those prior surveillances were
positive (i.e. a genuine previous-negative -> current-positive conversion).
}
}
