% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{get_non_single_patient_clusters}
\alias{get_non_single_patient_clusters}
\title{Get clusters containing more than one patient}
\usage{
get_non_single_patient_clusters(isolate_lookup)
}
\arguments{
\item{isolate_lookup}{A lookup table for isolates and their cluster assignments with other
relevant epidemiological information. See \code{\link[=get_isolate_lookup]{get_isolate_lookup()}}.}
}
\value{
A numeric vector of cluster IDs that contain isolates from more than one patient.
}
\description{
Get clusters containing more than one patient
}
