% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_algorithms.R
\name{get_tn_clusters_MSV_SVst_index_first}
\alias{get_tn_clusters_MSV_SVst_index_first}
\title{Get Transmission Clusters}
\usage{
get_tn_clusters_MSV_SVst_index_first(
  dna_aln,
  snp_dist,
  ip_seqs,
  ip_pt_seqs,
  seq2pt,
  dates,
  pars = FALSE,
  tree = NULL
)
}
\arguments{
\item{dna_aln}{A DNA object for sequences of interest.}

\item{snp_dist}{A matrix of SNP distances between isolates.}

\item{ip_seqs}{A vector of sequence IDs which correspond to intake patient sequences
presumed to be imported.}

\item{ip_pt_seqs}{A vector of sequence IDs which correspond to intake-positive patients.}

\item{seq2pt}{A named vector linking sequence IDs to patient IDs.}

\item{dates}{A vector of isolate dates named by sequence IDs.}

\item{pars}{Logical, if TRUE then use parsimony tree instead of the default neighbor joining tree.}

\item{tree}{A tree to use if provided instead of the default neighbor joining tree.}
}
\value{
A vector indicating the cluster that each isolate belongs to.
}
\description{
This function identifies transmission clusters based on the number of shared variants.
}
\details{
Clustering is performed to identify the maximal clusters containing a single intake-positive patient
that occurs before all cluster converts. The clustering metric is the number of shared variants, and clusters
can have multiple intake-positive patients if they share an identical number of variants with other cluster
members or intake-positive patients occur after converts. This clustering also requires that clusters be
defined by at least one shared variant that other isolates don't have.
}
