% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_algorithms.R
\name{get_tn_clusters_snp_thresh_tree}
\alias{get_tn_clusters_snp_thresh_tree}
\title{Perform clustering of isolates using a hard SNP distance cutoff.}
\usage{
get_tn_clusters_snp_thresh_tree(tree, snp_dist, snp_thresh)
}
\arguments{
\item{tree}{A phylogenetic tree object of class \code{phylo} constructed from the DNA alignment.}

\item{snp_dist}{A matrix of SNP distances between isolates constructed using a model of DNA evolution.
See \code{\link{get_snp_dist_matrix}} for a useful function to generate this.}

\item{snp_thresh}{A threshold for defining clusters.}
}
\value{
A numeric vector indicating the cluster that each isolate belongs to.
}
\description{
This function uses a hard SNP distance cutoff to define clusters. It first cuts the tree at the specified
threshold and then assigns each cluster to the best matching subtree. The best matching subtree is determined by
counting how many sequences in the cluster are in each subtree. If a cluster has more than one sequence, it is
assigned to the smallest matching subtree.
}
