% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitraceR-package.R
\docType{package}
\name{hospitraceR-package}
\alias{hospitraceR}
\alias{hospitraceR-package}
\title{hospitraceR: Detecting and analyzing nosocomial transmission clusters using genomic epidemiology}
\description{
Detects transmission clusters of healthcare-associated pathogens by combining whole-genome single-nucleotide-variant data with patient location traces, surveillance cultures and admission status. Also provides tools to compare cluster assignments, to test spatial and temporal patient overlap within clusters by permutation, and to categorize the inferred transmission role of each patient.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://theabhirath.github.io/hospitraceR}
  \item \url{https://github.com/theabhirath/hospitraceR}
  \item Report bugs at \url{https://github.com/theabhirath/hospitraceR/issues}
}

}
\author{
\strong{Maintainer}: Abhirath Anand \email{abhirath@umich.edu} (\href{https://orcid.org/0009-0008-6985-9266}{ORCID})

Authors:
\itemize{
  \item Abhirath Anand \email{abhirath@umich.edu} (\href{https://orcid.org/0009-0008-6985-9266}{ORCID})
}

}
\keyword{internal}
