% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{intra_cluster_genetic_distances}
\alias{intra_cluster_genetic_distances}
\title{Compute Genetic Distances}
\usage{
intra_cluster_genetic_distances(seqs, snp_dist)
}
\arguments{
\item{seqs}{Vector of sequence IDs in the cluster}

\item{snp_dist}{Matrix of SNP distances between isolates}
}
\value{
List with genetic distance properties
}
\description{
Calculate genetic distance statistics within a cluster
}
