% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_overlap_stats.R
\name{percentage_converts_with_overlap}
\alias{percentage_converts_with_overlap}
\title{Calculate the percentage of converts with overlap}
\usage{
percentage_converts_with_overlap(cluster_overlap_df, isolate_lookup)
}
\arguments{
\item{cluster_overlap_df}{A data frame with overlap information for isolate pairs. For more
information, see \code{\link{cluster_isolate_overlap}}.}

\item{isolate_lookup}{A lookup table for isolates and their clusters assignments which has
other relevant epidemiological information. For more information, see
\code{\link{get_isolate_lookup}}.}
}
\value{
A named vector with the percentage of converts with overlap for each cluster.
}
\description{
This function calculates the percentage of converts with overlap for each cluster.
}
