% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{plot_cluster_patient_trace}
\alias{plot_cluster_patient_trace}
\title{Plots a trace heatmap for all patients for a cluster.}
\usage{
plot_cluster_patient_trace(
  tree,
  cluster_seqs,
  dates,
  ip_seqs,
  seq2pt,
  trace_data,
  trace_colors = paletteer_d("ggthemes::Classic_Cyclic"),
  surv_colors = c("darkblue", "red", "yellow")
)
}
\arguments{
\item{tree}{A phylogenetic tree object of class \code{phylo}.}

\item{cluster_seqs}{A vector of sequence IDs which correspond to the cluster.}

\item{dates}{A named vector of isolate dates by sequence ID.}

\item{ip_seqs}{A vector of sequence IDs which correspond to intake patient sequences presumed to be imported.}

\item{seq2pt}{A named vector mapping sequence IDs to patient IDs.}

\item{trace_data}{A data frame or matrix representing daily patient trace (rows) by patient ID (columns).}

\item{trace_colors}{A vector of colors for the trace, one for each trace value.}

\item{surv_colors}{A vector of two colors for the surveillance, negative and positive.}
}
\value{
A \code{ggplot} object with the trace heatmap for all patients for a cluster.
}
\description{
Plots a trace heatmap for all patients for a cluster.
}
