% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{plot_clusters_phylo}
\alias{plot_clusters_phylo}
\title{Maps clusters onto a phylogenetic tree and visualizes them}
\usage{
plot_clusters_phylo(clusters, tree)
}
\arguments{
\item{clusters}{A vector named by sequence IDs with values indicating the cluster (subtree) each sequence belongs to.}

\item{tree}{A phylogenetic tree object of class \code{phylo}.}
}
\value{
A \code{ggtree} object with clusters visualized on the tree.
}
\description{
Maps clusters onto a phylogenetic tree and visualizes them
}
\author{
\href{mailto:aryansin@umich.edu}{Aryan Singh}
}
