% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_input_data.R
\name{read_in_seq_aln}
\alias{read_in_seq_aln}
\title{Read in Sequence Alignment}
\usage{
read_in_seq_aln(aln_file)
}
\arguments{
\item{aln_file}{A character string specifying the path to the alignment file in FASTA format.}
}
\value{
A DNA object with cleaned sequence names.
}
\description{
This function reads in a sequence alignment file and processes the sequence names.
}
\details{
The function performs the following steps:
\itemize{
\item Reads the DNA sequences from the specified FASTA file.
\item Cleans up the sequence names by removing specific substrings and patterns.
\item Replaces certain sequence names with predefined values.
\item Prints the cleaned sequence names.
\item Assigns the cleaned sequence names to the DNA object.
}
}
\keyword{internal}
