% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{remove_single_patient_clusters}
\alias{remove_single_patient_clusters}
\title{Remove clusters with only one patient from a vector of cluster assignments.}
\usage{
remove_single_patient_clusters(isolate_lookup)
}
\arguments{
\item{isolate_lookup}{A lookup table for isolates and their clusters assignments which has
other relevant epidemiological information. For more information, see
\code{\link{get_isolate_lookup}}.}
}
\value{
A lookup table for isolates and their clusters assignments with the single patient clusters removed.
}
\description{
This function removes clusters with only one patient from a vector of cluster assignments.
}
