% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transclust-package.R
\docType{package}
\name{transclust-package}
\alias{transclust}
\alias{transclust-package}
\title{transclust: An R package for transmission cluster analysis for pathogens}
\description{
An R package for transmission cluster analysis for pathogens.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://theabhirath.github.io/transclust}
  \item \url{https://github.com/theabhirath/transclust}
  \item Report bugs at \url{https://github.com/theabhirath/transclust/issues}
}

}
\author{
\strong{Maintainer}: Abhirath Anand \email{abhirath@umich.edu} (\href{https://orcid.org/0009-0008-6985-9266}{ORCID})

}
\keyword{internal}
