% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_overlap.R
\name{cluster_isolate_overlap}
\alias{cluster_isolate_overlap}
\title{Calculate isolate-isolate overlap for clusters}
\usage{
cluster_isolate_overlap(isolate_lookup, iso_overlap_df)
}
\arguments{
\item{isolate_lookup}{A lookup table for isolates and their clusters assignments which has
other relevant epidemiological information. For more information, see
\code{\link[=get_isolate_lookup]{get_isolate_lookup()}}.}

\item{overlap_df}{A data frame with overlap information for isolate pairs. For more information, see
\code{\link[=isolate_isolate_overlap]{isolate_isolate_overlap()}}.}
}
\value{
A data frame with columns: cluster, isolate_id, overlap.
}
\description{
This function considers every isolate in a cluster as a recipient, and returns if there is
any overlap when other isolates in the cluster are the donors.
}
