% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{compare_clusters}
\alias{compare_clusters}
\title{Compare Clusters}
\usage{
compare_clusters(clusters1, clusters2, prefix)
}
\arguments{
\item{clusters1}{A vector named by sequence IDs with values being subtrees defining the cluster.}

\item{clusters2}{A vector named by the same sequence IDs as clusters1 with values being subtrees
defining the cluster.}

\item{prefix}{Prefix to use for naming figure output files.}
}
\value{
A comparison of the content of clusters.
}
\description{
Compares the content of clusters created with two different methods.
}
\examples{
\dontrun{
compare_clusters(clusters1, clusters2, "prefix")
}

}
