% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeDefiningVariants}
\alias{computeDefiningVariants}
\title{Compute the defining variants for each subtree}
\usage{
computeDefiningVariants(dna_aln, isolate_names, subtrees)
}
\arguments{
\item{dna_aln}{A matrix of DNA sequences (\code{CharacterMatrix}).}

\item{isolate_names}{A vector of isolate names (\code{CharacterVector}).}

\item{subtrees}{A list of subtrees (\code{phylo} objects).}
}
\value{
An \code{IntegerVector} of integers representing the number of defining
variants for each subtree.
}
\description{
Compute the defining variants for each subtree
}
