% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{get_non_single_patient_clusters}
\alias{get_non_single_patient_clusters}
\title{Get non-single patient clusters from a vector of cluster assignments.}
\usage{
get_non_single_patient_clusters(isolate_lookup)
}
\arguments{
\item{isolate_lookup}{A lookup table for isolates and their clusters assignments which has
other relevant epidemiological information. For more information, see
\code{\link[=get_isolate_lookup]{get_isolate_lookup()}}.}
}
\value{
A numeric vector of cluster assignments with the non-single patient clusters.
}
\description{
This function gets non-single patient clusters from a vector of cluster assignments.
}
