% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_phylo_tree}
\alias{get_phylo_tree}
\title{Get phylogenetic tree using neighbor-joining or maximum parsimony method.}
\usage{
get_phylo_tree(dna_aln, snp_dist, method = c("nj", "pars"))
}
\arguments{
\item{dna_aln}{A DNA alignment object.}

\item{snp_dist}{A SNP distance matrix.}

\item{method}{A string indicating the method to use for tree construction. Options are
"nj" (neighbor-joining) or "pars" (maximum parsimony).}
}
\value{
An object of class \code{phylo} representing the phylogenetic tree.
}
\description{
This function constructs a neighbor-joining or maximum parsimony phylogenetic
tree from a DNA alignment object and a SNP distance matrix.
}
