% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_snp_dist_matrix}
\alias{get_snp_dist_matrix}
\title{Get SNP distance matrix}
\usage{
get_snp_dist_matrix(dna_aln, core = TRUE)
}
\arguments{
\item{dna_aln}{A DNA alignment object.}

\item{core}{Logical, if TRUE, return the core SNP distance matrix, otherwise
return the full SNP distance matrix.}
}
\value{
A SNP distance matrix.
}
\description{
Get SNP distance matrix from DNA alignment constructed using a model of DNA evolution.
}
