% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_algorithms.R
\name{get_tn_clusters_snp_thresh}
\alias{get_tn_clusters_snp_thresh}
\title{get_tn_clusters_snpthresh}
\usage{
get_tn_clusters_snp_thresh(dna_aln, snp_dist, snp_thresh)
}
\arguments{
\item{dna_aln}{A DNA object for sequences of interest.}

\item{snp_dist}{A matrix of SNP distances between isolates.}

\item{snp_thresh}{A threshold for defining clusters.}
}
\value{
A vector indicating the cluster that each isolate belongs to.
}
\description{
Performs clustering of isolates using a hard SNP distance cutoff.
}
