% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_algorithms.R
\name{get_tn_clusters_snp_thresh}
\alias{get_tn_clusters_snp_thresh}
\title{Perform clustering of isolates using a hard SNP distance cutoff.}
\usage{
get_tn_clusters_snp_thresh(snp_dist, snp_thresh)
}
\arguments{
\item{snp_dist}{A matrix of SNP distances between isolates constructed using a model of DNA evolution.
See \code{\link{get_snp_dist_matrix}} for a useful function to generate this.}

\item{snp_thresh}{A threshold for defining clusters.}
}
\value{
A numeric vector indicating the cluster that each isolate belongs to.
}
\description{
This function uses a hard SNP distance cutoff to define clusters naïvely, without using
the phylogenetic tree.
}
