% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitraceR-package.R
\docType{package}
\name{hospitraceR-package}
\alias{hospitraceR}
\alias{hospitraceR-package}
\title{hospitraceR: An R package for transmission cluster analysis for pathogens}
\description{
An R package for transmission cluster analysis for pathogens.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://theabhirath.github.io/hospitraceR}
  \item \url{https://github.com/theabhirath/hospitraceR}
  \item Report bugs at \url{https://github.com/theabhirath/hospitraceR/issues}
}

}
\author{
\strong{Maintainer}: Abhirath Anand \email{abhirath@umich.edu} (\href{https://orcid.org/0009-0008-6985-9266}{ORCID})

Authors:
\itemize{
  \item Abhirath Anand \email{abhirath@umich.edu} (\href{https://orcid.org/0009-0008-6985-9266}{ORCID})
}

}
\keyword{internal}
