% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{intra_cluster_duration_metrics}
\alias{intra_cluster_duration_metrics}
\title{Compute intra-cluster duration metrics}
\usage{
intra_cluster_duration_metrics(seqs, seq2pt, dates)
}
\arguments{
\item{seqs}{Vector of sequence IDs in the cluster}

\item{seq2pt}{Named vector mapping sequence IDs to patient IDs}

\item{dates}{Vector of isolate dates named by sequence IDs}
}
\value{
Vector with cluster duration properties
}
\description{
Calculate temporal metrics for spread of a single cluster.
}
