% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_overlap.R
\name{isolate_isolate_overlap}
\alias{isolate_isolate_overlap}
\title{Calculate isolate-isolate overlap}
\usage{
isolate_isolate_overlap(isolate_lookup, trace_mat)
}
\arguments{
\item{isolate_lookup}{A lookup table for isolates and their clusters assignments which has
other relevant epidemiological information. For more information, see
\code{\link{get_isolate_lookup}}.}

\item{trace_mat}{A matrix with rows representing days and columns representing patients.}
}
\value{
A data frame with columns: iso_donor, iso_recipient, overlap_days.
}
\description{
This function calculates the overlap between isolates in a given trace matrix.
}
