% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{plot_clusters_phylo}
\alias{plot_clusters_phylo}
\title{Maps clusters onto a phylogenetic tree and visualizes them}
\usage{
plot_clusters_phylo(
  tree,
  clusters,
  seq2pt = NULL,
  dna_var = NULL,
  patient_label = FALSE,
  convert_status = FALSE,
  ip_seqs = NULL,
  dates = NULL,
  pt_trace = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree object of class \code{phylo}.}

\item{clusters}{A vector named by sequence IDs with values indicating the cluster (subtree) each sequence belongs to.}

\item{seq2pt}{OPTIONAL. A named vector mapping sequence IDs to patient IDs. Default is NULL – no patient labels will
be added to the tree.}

\item{patient_label}{OPTIONAL. A logical value indicating whether to add patient labels to the tree. Default is FALSE
– no patient labels will be added to the tree.}
}
\value{
A \code{ggtree} object with clusters visualized on the tree.
}
\description{
Maps clusters onto a phylogenetic tree and visualizes them
}
\author{
\href{mailto:aryansin@umich.edu}{Aryan Singh} -
ORCID ID: \href{https://orcid.org/0000-0002-1850-5598}{0000-0002-1850-5598}
}
