% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{remap_cluster_values}
\alias{remap_cluster_values}
\title{Remap cluster values: each unique value (including each special value) gets its own number
in order of appearance.}
\usage{
remap_cluster_values(x, special_val = 0)
}
\arguments{
\item{x}{A numeric vector of cluster assignments.}

\item{special_val}{A value to treat as special i.e. it will get its own number in order of appearance.
This is useful for values that are not part of the cluster assignment, such as
singleton clusters.}
}
\value{
A numeric vector of remapped cluster assignments.
}
\description{
This function remaps cluster values so that each unique value (including each special value)
gets its own number in order of appearance.
}
\keyword{internal}
