% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_singleton_clusters}
\alias{remove_singleton_clusters}
\title{Remove singleton clusters from a vector of cluster assignments.}
\usage{
remove_singleton_clusters(clusters)
}
\arguments{
\item{clusters}{A numeric vector of cluster assignments.}
}
\description{
This function removes clusters with only one sequence from a vector of cluster assignments.
}
